allows segmentation for vqa and EL workspaces
