+++ ui/components/datasetItemWorkspace/src/components/DatasetItemViewer/DatasetItemViewer.svelte
@@ -7,10 +7,14 @@ License: CECILL-C
 <script lang="ts">
   // Imports
   import { Loader2Icon } from "lucide-svelte";
+  import * as ort from "onnxruntime-web";
+  import { afterUpdate } from "svelte";
 
   import { WorkspaceType, type DatasetItem } from "@pixano/core";
   import type { InteractiveImageSegmenterOutput } from "@pixano/models";
 
+  import { loadViewEmbeddings } from "../../lib/api/modelsApi";
+  import { modelsUiStore } from "../../lib/stores/datasetItemWorkspaceStores";
   import ThreeDimensionsViewer from "./3DViewer.svelte";
   import EntityLinkingViewer from "./EntityLinkingViewer.svelte";
   import ImageViewer from "./ImageViewer.svelte";
@@ -20,6 +24,40 @@ License: CECILL-C
   export let selectedItem: DatasetItem;
   export let currentAnn: InteractiveImageSegmenterOutput | null = null;
   export let isLoading: boolean;
+
+  let embeddings: Record<string, ort.Tensor> = {};
+
+  afterUpdate(() => {
+    if (
+      selectedItem &&
+      $modelsUiStore.yetToLoadEmbedding &&
+      $modelsUiStore.selectedModelName !== "" &&
+      $modelsUiStore.selectedTableName !== ""
+    ) {
+      modelsUiStore.update((store) => ({ ...store, yetToLoadEmbedding: false }));
+      loadViewEmbeddings(
+        selectedItem.item.id,
+        $modelsUiStore.selectedTableName,
+        selectedItem.ui.datasetId,
+      )
+        .then((results) => {
+          embeddings = results;
+          modelsUiStore.update((store) => ({
+            ...store,
+            currentModalOpen: "none",
+          }));
+        })
+        .catch((err) => {
+          modelsUiStore.update((store) => ({
+            ...store,
+            selectedTableName: "",
+            currentModalOpen: "noEmbeddings",
+          }));
+          console.error("cannot load Embeddings", err);
+          embeddings = {};
+        });
+    }
+  });
 </script>
 
 <div class="max-h-[calc(100vh-80px)] max-w-full bg-slate-800">
@@ -30,11 +68,11 @@ License: CECILL-C
   {:else if selectedItem.ui.type === WorkspaceType.VIDEO}
     <VideoViewer {selectedItem} bind:currentAnn />
   {:else if selectedItem.ui.type === WorkspaceType.IMAGE_VQA}
-    <VqaViewer {selectedItem} bind:currentAnn />
+    <VqaViewer {selectedItem} {embeddings} bind:currentAnn />
   {:else if selectedItem.ui.type === WorkspaceType.IMAGE_TEXT_ENTITY_LINKING}
-    <EntityLinkingViewer {selectedItem} bind:currentAnn />
+    <EntityLinkingViewer {selectedItem} {embeddings} bind:currentAnn />
   {:else if selectedItem.ui.type === WorkspaceType.IMAGE || !selectedItem.ui.type}
-    <ImageViewer {selectedItem} bind:currentAnn />
+    <ImageViewer {selectedItem} {embeddings} bind:currentAnn />
   {:else if selectedItem.ui.type === WorkspaceType.PCL_3D}
     <ThreeDimensionsViewer />
   {/if}
