+++ ui/components/datasetItemWorkspace/src/components/DatasetItemViewer/EntityLinkingViewer.svelte
@@ -8,6 +8,7 @@ License: CECILL-C
   // Imports
   import { Image as ImageJS } from "image-js";
   import { Loader2Icon } from "lucide-svelte";
+  import * as ort from "onnxruntime-web";
 
   import { Canvas2D } from "@pixano/canvas2d";
   import { DatasetItem, Image, Message, type ImagesPerView, type SaveItem } from "@pixano/core";
@@ -27,6 +28,7 @@ License: CECILL-C
     itemMasks,
     itemMetas,
     messages,
+    modelsUiStore,
     newShape,
     preAnnotationIsActive,
     saveData,
@@ -38,6 +40,7 @@ License: CECILL-C
   // Attributes
   export let selectedItem: DatasetItem;
   export let currentAnn: InteractiveImageSegmenterOutput | null = null;
+  export let embeddings: Record<string, ort.Tensor> = {};
 
   // Images per view type
   let imagesPerView: ImagesPerView = {};
@@ -110,6 +113,7 @@ License: CECILL-C
   const updateImages = async (): Promise<void> => {
     if (selectedItem.views) {
       loaded = false;
+      modelsUiStore.update((store) => ({ ...store, yetToLoadEmbedding: true }));
       imagesPerView = await loadImages(selectedItem.views as Record<string, Image>);
       loaded = true;
     }
@@ -173,6 +177,7 @@ License: CECILL-C
       masks={$itemMasks}
       keypoints={$itemKeypoints}
       selectedKeypointTemplate={templates.find((t) => t.template_id === $selectedKeypointsTemplate)}
+      {embeddings}
       {filters}
       imageSmoothing={$imageSmoothing}
       bind:selectedTool={$selectedTool}
